fixed my posts

+++ client/js/my-posts.js
@@ -1,4 +1,3 @@
-
 let data;
 
 function renderData() {
@@ -8,39 +7,51 @@ function renderData() {
 }
 
 async function deletePost(data) {
-    const response = await fetch(`/posts/delete/${data.id}`, {
-        method: "DELETE",
-        headers: {
-            'Content-Type': 'application/json',
-        },
-    })
-    const dataRes = await response.json()
-    console.log(dataRes)
+    console.log(data.id)
+    if (confirm("ARE YOU SURE YOU WANT TO DELETE!") === true) {
+        const response = await fetch(`/posts/delete/${data.id}`, {
+            method: "DELETE",
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        })
+        const dataRes = await response.json()
+        const stat = await response.status
+
+        if (stat === 200) {
+            renderPage()
+            alert("Successfully deleted post")
+        }
+        console.log(dataRes)
+
+    }
 }
 
 async function updatePost(data) {
-    const response = await fetch(`/posts/update/${data.id}`, {
-        method: "PUT",
-        headers: {
-            'Content-Type': 'application/json',
-        },
-    })
-    const dataRes = await response.json()
-    console.log(dataRes)
+    window.localStorage.setItem('update-id', data.id);
+    window.location.href = 'update-post.html'
+    // const response = await fetch(`/posts/update/${data.id}`, {
+    //     method: "PUT",
+    //     headers: {
+    //         'Content-Type': 'application/json',
+    //     },
+    // })
+    // const dataRes = await response.json()
+    // console.log(dataRes)
 }
 
 const addPost = (data) => {
     const temp = document.getElementsByTagName("template")[0];
     const clon = temp.content.cloneNode(true);
     clon.getElementById('post-template-title').innerText = data.name;
-    clon.getElementById('post-template-cuisine').innerText = data.cuisine;
+    clon.getElementById('post-template-cuisine').innerText = data.cuisine === "none" ? "" : data.cuisine;
     clon.getElementById('post-template-description').innerText = data.description;
-    clon.getElementById('post-template-img').src = data.img;
-    clon.getElementById('post-template-author').innerText = 'By: User' + data.user_id;
+    clon.getElementById('post-template-img').src = 'data:image/png;base64,' + data.image;
+    clon.getElementById('post-template-author').innerText = 'By: User#' + data.userid;
     const ingredientsList = clon.getElementById('post-template-ingredients');
     const directionsList = clon.getElementById('post-template-directions');
-    const ingredients = data.ingredients;
-    const directions = data.directions;
+    const ingredients = data.ingredients.split(', ');
+    const directions = data.directions.split('.');
     ingredients.forEach(ingredient => {
         const li = document.createElement('li');
         li.innerText = ingredient;
@@ -61,18 +72,23 @@ const addPost = (data) => {
 
 const posts = document.getElementById("posts-main");
 
-const res = await fetch('/posts/get/User/2', {
-    method: "GET",
-    headers: {
-        'Content-Type': 'application/json',
-    },
-})
+async function renderPage() {
+    posts.innerHTML = ""
+    const res = await fetch('/posts/get/user', {
+        method: "GET",
+        headers: {
+            'Content-Type': 'application/json',
+        },
+    })
 
-data = await res.json()
-const status = await res.status
+    data = await res.json()
+    const status = await res.status
 
 
 
-if (status === 200) {
-    renderData()
+    if (status === 200) {
+        renderData()
+    }
 }
+
+renderPage()
